Move javascript translations from wirecloud.platform to wirecloud.commons
